LayoutTests/http/tests/site-isolation/resources/permissions-policy-expected.txt

CONSOLE MESSAGE: fail
FAIL: Timed out waiting for notifyDone to be called

Verifies permissions policy is set correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
